requirements.txt
accelerate>=0.23.0
autocrop>=1.3.0
awscli>=1.29.41
bitsandbytes>=0.41.0
diffusers>=0.21.1
matplotlib>=3.7.2
pillow>=9.4.0
torch>=2.0.0
torchvision>=0.15.2
sagemaker>=2.183.0
tensorboard>=2.14.0
tqdm>=4.65.0
transformers>=4.33.2
wandb>=0.15.11
xformers>=0.0.19